Support using 'webtop' as the system partition
The Motorola Droid 4's system partition is ~600Mb, which is too small
to be able to contain modern ROMs based off of Android 7 and a Gapps
package. This change adds an extra 'slot' to the recovery that allows
the repurposing of the otherwise unused 'webtop' partition to be used
in place of the stock system parition.